+++ src/android_helper/jni_onload.cc
@@ -26,7 +26,7 @@ extern "C" jint JNIEXPORT JNICALL JNI_OnLoad(JavaVM* jvm, void* reserved) {
   if (ret < 0)
     return -1;
 
-  RTC_CHECK(rtc::InitializeSSL()) << "Failed to InitializeSSL()";
+  RTC_CHECK(webrtc::InitializeSSL()) << "Failed to InitializeSSL()";
   webrtc::InitClassLoader(GetEnv());
   //LoadGlobalClassReferenceHolder();
   //unity_plugin::LoadGlobalClassReferenceHolder();
@@ -38,7 +38,7 @@ extern "C" jint JNIEXPORT JNICALL JNI_OnLoad(JavaVM* jvm, void* reserved) {
 extern "C" void JNIEXPORT JNICALL JNI_OnUnLoad(JavaVM* jvm, void* reserved) {
   //FreeGlobalClassReferenceHolder();
   //unity_plugin::FreeGlobalClassReferenceHolder();
-  RTC_CHECK(rtc::CleanupSSL()) << "Failed to CleanupSSL()";
+  RTC_CHECK(webrtc::CleanupSSL()) << "Failed to CleanupSSL()";
 
   RTC_LOG(LS_INFO) << "JNI_OnUnLoad";
 }