+++ src/unity_camera_capturer.h
@@ -29,7 +29,7 @@ struct UnityCameraCapturerConfig : sora::ScalableVideoTrackSourceConfig {
 };
 
 class UnityCameraCapturer : public sora::ScalableVideoTrackSource,
-                            public rtc::VideoSinkInterface<webrtc::VideoFrame> {
+                            public webrtc::VideoSinkInterface<webrtc::VideoFrame> {
   webrtc::Clock* clock_ = webrtc::Clock::GetRealTimeClock();
 
   struct Impl {
@@ -38,7 +38,7 @@ class UnityCameraCapturer : public sora::ScalableVideoTrackSource,
                       void* camera_texture,
                       int width,
                       int height) = 0;
-    virtual rtc::scoped_refptr<webrtc::I420Buffer> Capture() = 0;
+    virtual webrtc::scoped_refptr<webrtc::I420Buffer> Capture() = 0;
   };
 
 #ifdef SORA_UNITY_SDK_WINDOWS
@@ -54,7 +54,7 @@ class UnityCameraCapturer : public sora::ScalableVideoTrackSource,
               void* camera_texture,
               int width,
               int height) override;
-    rtc::scoped_refptr<webrtc::I420Buffer> Capture() override;
+    webrtc::scoped_refptr<webrtc::I420Buffer> Capture() override;
   };
 #endif
 
@@ -71,7 +71,7 @@ class UnityCameraCapturer : public sora::ScalableVideoTrackSource,
               void* camera_texture,
               int width,
               int height) override;
-    rtc::scoped_refptr<webrtc::I420Buffer> Capture() override;
+    webrtc::scoped_refptr<webrtc::I420Buffer> Capture() override;
   };
 #endif
 
@@ -91,7 +91,7 @@ class UnityCameraCapturer : public sora::ScalableVideoTrackSource,
               void* camera_texture,
               int width,
               int height) override;
-    rtc::scoped_refptr<webrtc::I420Buffer> Capture() override;
+    webrtc::scoped_refptr<webrtc::I420Buffer> Capture() override;
   };
 #endif
 
@@ -110,7 +110,7 @@ class UnityCameraCapturer : public sora::ScalableVideoTrackSource,
               void* camera_texture,
               int width,
               int height) override;
-    rtc::scoped_refptr<webrtc::I420Buffer> Capture() override;
+    webrtc::scoped_refptr<webrtc::I420Buffer> Capture() override;
   };
 #endif
 
@@ -119,7 +119,7 @@ class UnityCameraCapturer : public sora::ScalableVideoTrackSource,
   bool stopped_ = false;
 
  public:
-  static rtc::scoped_refptr<UnityCameraCapturer> Create(
+  static webrtc::scoped_refptr<UnityCameraCapturer> Create(
       const UnityCameraCapturerConfig& config);
 
   UnityCameraCapturer(const UnityCameraCapturerConfig& config);